--- Source/TestProject/Private/Item/EItemQuality.cpp
@@ -1,1 +1,0 @@
-﻿#include "Item/EItemQuality.h"
