--- Source/TestProject/Private/Item/EItemType.cpp
@@ -1,1 +1,0 @@
-﻿#include "Item/EItemType.h"
