+++ Source/TestProject/Public/Inventory/EItemQuality.h
